Did a lot of refactoring, and just added prop-types because Stephen told me to.
