+++ client/src/components/sections/recipe/components/Recipe.js
@@ -1,8 +1,8 @@
 /* Recipe LANDING PAGE */
 
 import React, { Component } from "react"
-import Sidebar from "../../components/Sidebar"
-import Mininavbar from "../../components/Mininavbar"
+import Sidebar from "../../../../components/Sidebar"
+import Mininavbar from "../../../../components/Mininavbar"
 
 import "../css/Recipe.css"
