+++ client/src/components/sections/restaurant/components/Restaurant.js
@@ -1,6 +1,6 @@
 import React, { Component } from "react"
-import Sidebar from "../../components/Sidebar"
-import Mininavbar from "../../components/Mininavbar"
+import Sidebar from "../../../../components/Sidebar"
+import Mininavbar from "../../../../components/Mininavbar"
 import "../css/Restaurant.css"
 
 class Restaurant extends Component{
